+++ dynastats/transactions/models.py
@@ -7,7 +7,7 @@ from rosters.models import Roster
 class Transaction(models.Model):
     league_id = models.ForeignKey(League, on_delete=models.CASCADE, db_index=True)
     transaction_id = models.CharField(max_length=80, primary_key=True)
-    leg = models.IntegerField()
+    leg = models.PositiveSmallIntegerField()
     created = models.DateTimeField()
     status_updated = models.DateTimeField()
     status = models.CharField(max_length=80)
